Reactor Content Update 2

+++ Content.Goobstation.Server/Power/_FarHorizons/Power/Generation/FissionGenerator/ReactorPartSystem.cs
@@ -26,7 +26,7 @@ public sealed class ReactorPartSystem : SharedReactorPartSystem
     /// <returns></returns>
     public GasMixture? ProcessGas(ReactorPartComponent reactorPart, Entity<NuclearReactorComponent> reactorEnt, AtmosDeviceUpdateEvent args, GasMixture inGas)
     {
-        if (reactorPart.RodType != (byte)ReactorPartComponent.RodTypes.GasChannel)
+        if (reactorPart.RodType != ReactorPartComponent.RodTypes.GasChannel)
             return null;
 
         GasMixture? ProcessedGas = null;
